Add files via upload
PROBLEM STATEMENT : A US-based housing company named Surprise Housing has decided to enter the Australian market. The company is looking at prospective properties to buy to enter the market . We are required to build a regression model using regularization in order to predict the actual value of the prospective properties and decide whether to invest in them or not.